@@ -60,64 +61,196 @@
     "cjs": "0.0.11",
     "cjs-loader": "^0.1.0",
     "concurrently": "^4.1.0",
-    "electron": "4.2.4",
     "electron-builder": "20.41.0",
     "eslint": "^5.16.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.13.0",
+    "node-sass-chokidar": "^1.3.5",
     "node-sass-tilde-importer": "^1.0.2",
     "standard": "^10.0.2"
   },
   "dependencies": {
-    "@blueprintjs/core": "3.1.0",
-    "@blueprintjs/datetime": "3.1.0",
-    "@blueprintjs/select": "3.1.0",
-    "axios": "^0.18.0",
-    "bfx-hf-ext-plugin-bitfinex": "git+https://github.com/bitfinexcom/bfx-hf-ext-plugin-bitfinex.git",
-    "bfx-hf-indicators": "bitfinexcom/bfx-hf-indicators",
-    "bfx-hf-models": "git+https://github.com/bitfinexcom/bfx-hf-models.git",
-    "bfx-hf-models-adapter-lowdb": "git+https://github.com/bitfinexcom/bfx-hf-models-adapter-lowdb.git",
-    "bfx-hf-server": "git+https://github.com/bitfinexcom/bfx-hf-server.git",
-    "bfx-hf-util": "bitfinexcom/bfx-hf-util",
-    "bignumber.js": "^7.2.1",
-    "bitfinex-api-node": "^2.0.2",
-    "child_process": "^1.0.2",
+    "@babel/core": "7.4.3",
+    "@fortawesome/fontawesome-free": "^5.9.0",
+    "@projectstorm/react-diagrams": "^5.3.2",
+    "@sentry/browser": "^5.5.0",
+    "@svgr/webpack": "4.1.0",
+    "@typescript-eslint/eslint-plugin": "1.6.0",
+    "@typescript-eslint/parser": "1.6.0",
+    "aes-js": "^3.1.2",
+    "babel-eslint": "10.0.1",
+    "babel-jest": "^24.8.0",
+    "babel-loader": "8.0.5",
+    "babel-plugin-named-asset-import": "^0.3.2",
+    "babel-preset-react-app": "^9.0.0",
+    "bfx-api-node-models": "^1.0.10",
+    "bfx-api-node-util": "^1.0.1",
+    "bfx-hf-algo": "github:f3rno/bfx-hf-algo",
+    "bfx-hf-backtest": "github:f3rno/bfx-hf-backtest",
+    "bfx-hf-indicators": "github:f3rno/bfx-hf-indicators",
+    "bfx-hf-strategy": "github:f3rno/bfx-hf-strategy",
+    "bfx-hf-util": "github:f3rno/bfx-hf-util",
+    "bfx-hf-ext-plugin-bitfinex": "github:f3rno/bfx-hf-ext-plugin-bitfinex",
+    "bfx-hf-models": "github:f3rno/bfx-hf-models",
+    "bfx-hf-models-adapter-lowdb": "github:f3rno/bfx-hf-models-adapter-lowdb",
+    "bignumber.js": "^9.0.0",
+    "bufferutil": "^4.0.1",
+    "camelcase": "^5.2.0",
+    "case-sensitive-paths-webpack-plugin": "2.2.0",
     "classnames": "^2.2.6",
-    "codemirror": "^5.39.2",
-    "color-scheme": "^1.0.1",
-    "cors": "^2.8.5",
-    "d3": "^5.4.0",
-    "dotenv": "^8.0.0",
-    "electron-is-dev": "^1.1.0",
-    "express": "^4.17.1",
-    "express-request-proxy": "^2.2.2",
+    "closest": "0.0.1",
+    "codemirror": "^5.48.0",
+    "connected-react-router": "^6.5.2",
+    "css-loader": "2.1.1",
+    "d3-format": "^1.3.2",
+    "d3-time-format": "^2.1.3",
+    "debug": "^4.1.1",
+    "dotenv": "6.2.0",
+    "dotenv-expand": "4.2.0",
+    "electron": "4.2.4",
+    "email-validator": "^2.0.4",
+    "eslint": "^5.16.0",
+    "eslint-config-react-app": "^4.0.1",
+    "eslint-loader": "2.1.2",
+    "eslint-plugin-flowtype": "2.50.1",
+    "eslint-plugin-import": "2.16.0",
+    "eslint-plugin-jsx-a11y": "6.2.1",
+    "eslint-plugin-react": "7.12.4",
+    "eslint-plugin-react-hooks": "^1.6.1",
+    "file-loader": "3.0.1",
+    "fs-extra": "7.0.1",
+    "history": "^4.9.0",
+    "html-webpack-plugin": "4.0.0-beta.5",
+    "identity-obj-proxy": "3.0.0",
     "intercept-stdout": "^0.1.2",
+    "is-wsl": "^1.1.0",
+    "jest": "24.7.1",
+    "jest-environment-jsdom-fourteen": "0.1.0",
+    "jest-resolve": "24.7.1",
+    "jest-watch-typeahead": "0.3.0",
+    "js-levenshtein": "^1.1.6",
     "lodash": "^4.17.11",
-    "node-sass-chokidar": "^1.3.0",
-    "nohup": "^0.1.0",
+    "mini-css-extract-plugin": "0.5.0",
+    "monaco-editor": "^0.17.1",
+    "monaco-editor-webpack-plugin": "^1.7.0",
+    "normalize.css": "^8.0.1",
+    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "pathfinding": "^0.4.18",
+    "paths-js": "^0.4.9",
+    "pnp-webpack-plugin": "1.2.1",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "postcss-normalize": "7.0.1",
+    "postcss-preset-env": "6.6.0",
+    "postcss-safe-parser": "4.0.1",
+    "pretty-checkbox": "^3.0.3",
     "prop-types": "^15.7.2",
-    "react": "^16.4.0",
-    "react-codemirror2": "^5.1.0",
-    "react-dom": "^16.4.0",
-    "react-modal": "^3.9.1",
+    "randomcolor": "^0.5.4",
+    "raw.macro": "^0.3.0",
+    "react": "^16.8.6",
+    "react-app-polyfill": "^1.0.1",
+    "react-codemirror2": "^6.0.0",
+    "react-custom-scrollbars": "^4.2.1",
+    "react-datepicker": "^2.7.0",
+    "react-dev-utils": "^9.0.1",
+    "react-dom": "^16.8.6",
+    "react-ga": "^2.6.0",
+    "react-grid-layout": "^0.16.6",
+    "react-markdown": "^4.1.0",
     "react-monaco-editor": "^0.26.2",
-    "react-notifications": "^1.4.3",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "react-scripts": "1.1.4",
-    "react-stockcharts": "^0.7.7",
-    "react-switch": "^5.0.0",
-    "react-transition-group": "^2.3.1",
-    "react-virtualized": "^9.19.1",
-    "redux": "^4.0.0",
-    "redux-saga": "^0.16.0",
-    "redux-thunk": "^2.3.0",
-    "socks-proxy-agent": "^4.0.2",
-    "url": "^0.11.0",
-    "wait-on": "^3.2.0"
+    "react-onclickoutside": "^6.8.0",
+    "react-redux": "^7.1.0",
+    "react-router": "^5.0.1",
+    "react-select": "^3.0.4",
+    "react-stockcharts": "^0.7.8",
+    "react-tooltip": "^3.10.0",
+    "react-virtualized": "^9.21.1",
+    "redux": "^4.0.1",
+    "redux-beacon": "^2.1.0",
+    "redux-beacon-matomo-tag-manager": "^1.0.0",
+    "redux-saga": "^1.0.4",
+    "resolve": "1.10.0",
+    "sass-loader": "7.1.0",
+    "scrypt-js": "^2.0.4",
+    "semver": "6.0.0",
+    "sha.js": "^2.4.11",
+    "style-loader": "0.23.1",
+    "terser-webpack-plugin": "1.2.3",
+    "thread-loader": "^2.1.2",
+    "ts-pnp": "1.1.2",
+    "url-loader": "1.1.2",
+    "utf-8-validate": "^5.0.2",
+    "webpack": "4.29.6",
+    "webpack-dev-server": "3.2.1",
+    "webpack-manifest-plugin": "2.0.4",
+    "workbox-webpack-plugin": "4.2.0",
+    "worker-loader": "^2.0.0"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx,ts,tsx}",
+      "!src/**/*.d.ts"
+    ],
+    "setupFiles": [
+      "react-app-polyfill/jsdom"
+    ],
+    "setupFilesAfterEnv": [],
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
+      "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
+    ],
+    "testEnvironment": "jest-environment-jsdom-fourteen",
+    "transform": {
+      "^.+\\.(js|jsx|ts|tsx)$": "<rootDir>/node_modules/babel-jest",
+      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
+      "^(?!.*\\.(js|jsx|ts|tsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
+    },
+    "transformIgnorePatterns": [
+      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$",
+      "^.+\\.module\\.(css|sass|scss)$"
+    ],
+    "modulePaths": [],
+    "moduleNameMapper": {
+      "^react-native$": "react-native-web",
+      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
+    },
+    "moduleFileExtensions": [
+      "web.js",
+      "js",
+      "web.ts",
+      "ts",
+      "web.tsx",
+      "tsx",
+      "json",
+      "web.jsx",
+      "jsx",
+      "node"
+    ],
+    "watchPlugins": [
+      "jest-watch-typeahead/filename",
+      "jest-watch-typeahead/testname"
+    ]
+  },
+  "babel": {
+    "presets": [
+      "react-app"
+    ]
   }
 }
