@@ -1,18 +1,13 @@
-import {
-  default as LabelAnnotation, helper,
+import LabelAnnotation, {
+  helper,
 } from 'react-stockcharts/lib/annotation/LabelAnnotation'
 
 import React from 'react'
 import CaretDownIcon from '../../../ui/Icons/CaretDownIcon'
-import { labelForStrategyTradeAnnotation } from '../Chart.helpers'
+import { labelForStrategyTradeAnnotation } from '../helpers'
 
-/**
- * Renders multiple caret-down icons, one for each order within the current candle
- * specified via 'datum'
- */
 export default class SellOrder extends LabelAnnotation {
   static propTypes = LabelAnnotation.propTypes
-
   static defaultProps = LabelAnnotation.defaultProps
 
   render() {
@@ -23,7 +18,7 @@ export default class SellOrder extends LabelAnnotation {
     const orders = mtsMap[datum.mts]
 
     return orders.map(o => (
-      <g className={className} key={o.mts}>
+      <g className={className} key={o.order_id}>
         <title>{labelForStrategyTradeAnnotation(o)}</title>
         <CaretDownIcon
           x={xPos}
