@@ -1,87 +1,129 @@
 import React from 'react'
-import { Spinner, Intent } from '@blueprintjs/core'
-import { Route, Redirect } from 'react-router-dom'
-import _isEmpty from 'lodash/isEmpty'
-import { NotificationContainer } from 'react-notifications'
+import { Route, Switch } from 'react-router'
 
-import APIComboDialog from '../APIComboDialog'
-import SideNavBar from '../../ui/SideNavBar'
-import StatusBar from '../../ui/StatusBar'
-import AlgoOrdersView from '../../pages/AlgoOrders'
-import SettingsView from '../../pages/Settings'
-import { propTypes } from './HFUI.props'
+import TradingPage from '../../pages/Trading'
+import StrategyEditorPage from '../../pages/StrategyEditor'
+import MarketDataPage from '../../pages/MarketData'
+
+import Navbar from '../Navbar'
+import LoginModal from '../LoginModal'
+import RegisterModal from '../RegisterModal'
+import ExchangeInfoBar from '../ExchangeInfoBar'
+import NotificationsSidebar from '../NotificationsSidebar'
+
+import { propTypes, defaultProps } from './HFUI.props'
+import './style.css'
 
 export default class HFUI extends React.Component {
   static propTypes = propTypes
+  static defaultProps = defaultProps
+
+  state = {
+    loginModalOpen: false,
+    registerModalOpen: false,
+  }
 
   constructor(props) {
     super(props)
 
-    this.onSubmitKeys = this.onSubmitKeys.bind(this)
+    this.onChangeMarket = this.onChangeMarket.bind(this)
+    this.onToggleLoginModal = this.onToggleLoginModal.bind(this)
+    this.onToggleRegisterModal = this.onToggleRegisterModal.bind(this)
   }
 
   componentDidMount() {
-    const { loadInitialSettings, loadAPIKey, cycleBFXConnection, getLastVersion } = this.props
+    const { loginWithAuthToken } = this.props
+    const cookies = document.cookie.split(';')
 
-    loadInitialSettings()
-    loadAPIKey()
-    cycleBFXConnection()
-    getLastVersion()
-    // keep trying to load api keys
-    const reloadKeys = () => {
-      const { apiKeyCombo = {} } = this.props
-      if (!_isEmpty(apiKeyCombo)) return
-      loadAPIKey()
-      setTimeout(reloadKeys, 5000)
-    }
-    reloadKeys()
+    cookies.forEach((c) => {
+      const cookie = c.split('=')
+
+      if (cookie[0].trim() === 'authToken') {
+        loginWithAuthToken(cookie[1])
+      }
+    })
   }
 
-  onSubmitKeys({ key, secret } = {}) {
-    const { submitAPIKey } = this.props
-    submitAPIKey({ key, secret })
+  onChangeMarket(option) {
+    const { saveActiveMarket } = this.props
+    saveActiveMarket(option.value)
+  }
+
+  onToggleLoginModal() {
+    this.setState(({ loginModalOpen }) => ({
+      loginModalOpen: !loginModalOpen,
+      registerModalOpen: false,
+    }))
+  }
+
+  onToggleRegisterModal() {
+    this.setState(({ registerModalOpen }) => ({
+      registerModalOpen: !registerModalOpen,
+      loginModalOpen: false,
+    }))
   }
 
   render() {
-    const { apiKeyCombo = {} } = this.props
+    const { loginModalOpen, registerModalOpen } = this.state
+    const { activeMarket } = this.props
 
-    if (_isEmpty(apiKeyCombo)) {
-      return (
-        <div className='bp3-dark hfui loading'>
-          <Spinner
-            intent={Intent.PRIMARY}
-            size={Spinner.SIZE_LARGE}
-          />
-        </div>
-      )
-    }
-
-    const { key, secret } = apiKeyCombo
-
-    if (!key || !secret) {
-      return (
-        <div className='bp3-dark hfui'>
-          <APIComboDialog
-            onSubmit={this.onSubmitKeys}
-          />
-        </div>
-      )
-    }
-
-    /*
-      <Route exact path='/index.html' component={DashboardView} />  used for an electron app
-    */
     return (
-      <div className='hfui'>
-        <div className='hfui_content__wrapper'>
-          <NotificationContainer />
-          <Route component={SideNavBar} />
-          <Redirect exact from='/' to='/algo-orders' />
-          <Route path='/algo-orders' component={AlgoOrdersView} />
-          <Route path='/settings' component={SettingsView} />
-        </div>
+      <div className='hfui-app'>
+        <Navbar
+          onLogin={this.onToggleLoginModal}
+          onRegister={this.onToggleRegisterModal}
+        />
 
-        <Route component={StatusBar} />
+        <ExchangeInfoBar
+          selectedMarket={activeMarket}
+          onChangeMarket={this.onChangeMarket}
+        />
+
+        <Switch>
+          <Route
+            exact
+            path='/'
+            render={() => (
+              <TradingPage
+                onRegister={this.onToggleRegisterModal}
+                onLogin={this.onToggleLoginModal}
+              />
+            )}
+          />
+
+          <Route
+            path='/strategy-editor'
+            render={() => (
+              <StrategyEditorPage onLogin={this.onToggleLoginModal} />
+            )}
+          />
+
+          <Route
+            path='/data'
+            render={() => (
+              <MarketDataPage
+                onRegister={this.onToggleRegisterModal}
+                onLogin={this.onToggleLoginModal}
+              />
+            )}
+          />
+        </Switch>
+
+        <NotificationsSidebar />
+
+        {loginModalOpen && (
+          <LoginModal
+            onClose={this.onToggleLoginModal}
+            onRegister={this.onToggleRegisterModal}
+          />
+        )}
+
+        {registerModalOpen && (
+          <RegisterModal
+            onClose={this.onToggleRegisterModal}
+            onLogin={this.onToggleLoginModal}
+          />
+        )}
       </div>
     )
   }
